+++ Settings.cpp
@@ -157,9 +157,9 @@ BOOL Settings_LoadBinaryValue(LPCTSTR sName, void * pData, int size)
 
         TCHAR ch = *buf;
         if (ch >= _T('0') && ch <= _T('9'))
-            v = ch - _T('0');
+            v = (BYTE)(ch - _T('0'));
         else if (ch >= _T('A') && ch <= _T('F'))
-            v = ch - _T('A') + 10;
+            v = (BYTE)(ch - _T('A') + 10);
         else  // Not hex
         {
             free(buffer);
@@ -171,9 +171,9 @@ BOOL Settings_LoadBinaryValue(LPCTSTR sName, void * pData, int size)
 
         ch = *buf;
         if (ch >= _T('0') && ch <= _T('9'))
-            v |= ch - _T('0');
+            v |= (BYTE)(ch - _T('0'));
         else if (ch >= _T('A') && ch <= _T('F'))
-            v |= ch - _T('A') + 10;
+            v |= (BYTE)(ch - _T('A') + 10);
         else  // Not hex
         {
             free(buffer);
